+++ src/components/EditorBox.jsx
@@ -20,7 +20,6 @@ function EditorBox({ onContentChange, value }) {
   const [content, setContent] = useState(null);
   const location = useLocation();
   useEffect(() => {
-    console.log(value === '');
     if (location.pathname.includes('edit')) {
       if (value !== '') {
         setIsCreating(false);
