+++ src/pages/LoginPage.jsx
@@ -1,129 +1,7 @@
-import { Link, useNavigate, useLocation } from 'react-router-dom';
-import { Box, Container, Typography, Input, Button } from '@mui/material';
-import { useState } from 'react';
-import { useAuth } from '../AuthenticationContext';
-import BASE_URL from '../config';
+import LogInForm from '../components/LogInForm';
 
 function LoginPage() {
-  const navigate = useNavigate();
-  const location = useLocation();
-  const { pathname } = location;
-  const isError = pathname.includes('error');
-  const [user] = useAuth();
-  const [isTyping, setIsTyping] = useState(false);
-
-  if (user) {
-    navigate('/dashboard');
-  }
-
-  const handleChange = () => {
-    setIsTyping(true);
-  };
-
-  return (
-    <Container className="Login">
-      <Typography
-        variant="h1"
-        textAlign="center"
-        sx={{
-          mt: '105px',
-        }}
-      >
-        로그인
-      </Typography>
-
-      <form className="form" action={`${BASE_URL}/api/login`} method="POST">
-        <Box className="flex-column">
-          <label className="labels" htmlFor="email" />
-          <Box className="inputForm">
-            <Input
-              onChange={handleChange}
-              required
-              fullWidth
-              type="email"
-              id="email"
-              name="email"
-              className="form-control"
-              placeholder="이메일을 입력해주세요."
-              sx={{
-                borderRadius: '16px',
-                mb: '16px',
-                mt: '57px',
-                backgroundColor: 'background.default',
-                border: '1px solid rgb(52, 52, 52)',
-              }}
-            />
-          </Box>
-        </Box>
-
-        <Box className="flex-column">
-          <label className="labels" htmlFor="password" />
-          <Box className="inputForm">
-            <Input
-              onChange={handleChange}
-              required
-              fullWidth
-              type="password"
-              id="password"
-              name="password"
-              className="form-control"
-              placeholder="비밀번호를 입력해주세요."
-              sx={{
-                borderRadius: '16px',
-                backgroundColor: 'background.default',
-                border: '1px solid rgb(52, 52, 52)',
-              }}
-            />
-          </Box>
-        </Box>
-        {isError && !isTyping ? (
-          <Box>
-            <Typography
-              textAlign="left"
-              fontSize="fontSizeMd"
-              fontWeight="fontWeightRegular"
-              sx={{
-                mt: '10px',
-                ml: '8px',
-                color: 'rgb(131, 36, 254)',
-              }}
-            >
-              *등록되지 않은 아이디이거나 비밀번호를 잘못 입력하셨습니다.
-            </Typography>
-          </Box>
-        ) : (
-          <Box>
-            <Typography />
-          </Box>
-        )}
-        <Button
-          fullWidth
-          type="submit"
-          className="button-submit"
-          sx={{
-            background: 'rgb(131, 36, 254)',
-            borderRadius: '16px',
-            height: '60px',
-            mt: '31px',
-          }}
-        >
-          <Typography>로그인</Typography>
-        </Button>
-      </form>
-      <Box
-        sx={{
-          display: 'flex',
-          justifyContent: 'center',
-          mt: '16px',
-        }}
-      >
-        <Typography sx={{ mr: '5px' }}>계정이 없으신가요?</Typography>
-        <Link to="/join" style={{ color: 'inherit', textDecoration: 'none' }}>
-          <Typography sx={{ color: 'rgb(131, 36, 254)' }}>회원가입</Typography>{' '}
-        </Link>
-      </Box>
-    </Container>
-  );
+  return <LogInForm />;
 }
 
 export default LoginPage;
